# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kantin Rumah Kayu API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/tenant/orders": {
            "get": {
                "tags": [
                    "Tenant Orders"
                ],
                "operationId": "6428ebca7b3971869c0781a685d798c6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created",
                                "preparing",
                                "ready_for_pickup",
                                "picked_up",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant order list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/v1/tenant/orders/{order}/status": {
            "patch": {
                "tags": [
                    "Tenant Orders"
                ],
                "operationId": "01e8cb7f1aefff4b51ef4bf3f07e2e63",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "preparing",
                                            "ready_for_pickup"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid transition"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tenant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Warung Nusantara"
                    },
                    "opens_at": {
                        "type": "string",
                        "example": "08:00:00"
                    },
                    "closes_at": {
                        "type": "string",
                        "example": "21:00:00"
                    }
                },
                "type": "object"
            },
            "Menu": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "example": "Nasi Goreng"
                    },
                    "price": {
                        "type": "integer",
                        "example": 25000
                    },
                    "category": {
                        "type": "string",
                        "example": "main"
                    },
                    "photo_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "menu_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "integer",
                        "example": 12000
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 24000
                    },
                    "menu": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Menu"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "integer"
                    },
                    "tenant": {
                        "$ref": "#/components/schemas/Tenant"
                    },
                    "type": {
                        "type": "string",
                        "example": "pickup"
                    },
                    "status": {
                        "type": "string",
                        "example": "ready_for_pickup"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "paid"
                    },
                    "total_price": {
                        "type": "integer",
                        "example": 24000
                    },
                    "paid_amount": {
                        "type": "integer",
                        "example": 24000
                    },
                    "picked_up_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_by_user": {
                        "type": "boolean",
                        "example": false
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}